Something went wrong with the last commit. I rewrote much of the things I did last week.

+++ benchmark/rubis/src/main/java/org/imdea/rubis/benchmark/stats/StatsCollector.java
@@ -63,7 +63,7 @@ public class StatsCollector {
         int ops = mReadOnlyLatencies.size() + mUpdateLatencies.size();
 
         printLine("OVERALL", "RunTime(ms)", time);
-        printLine("OVERALL", "Throughput(ops/sec)", ops / ((double) time / 1000));
+        printLine("OVERALL", "Throughput(ops/sec)", ops / (time / 1000.0));
         printLine("OVERALL", "Exceptions", mExceptions);
     }
